added endpoint to register, apply and verify promo codes

+++ izinga-qrcode-generator/src/main/kotlin/io/curiousoft/izinga/qrcodegenerator/promocodes/model/PromoCode.kt
@@ -5,7 +5,7 @@ import org.springframework.data.mongodb.core.mapping.Document
 import java.time.LocalDateTime
 
 @Document
-data class PromoCode(@Id val code: String, val expiryDate: LocalDateTime, val percentage: Int,
+data class PromoCode(@Id val code: String, val expiryDate: LocalDateTime, val percentage: Double,
                      val type: PromoType, val maxRedemption: Int, val storeId: String? = null)
 
 enum class PromoType {
